// src/lib/geocode/model/address.ts
/**
 * Service de géocodage
 *
 * Contact: contact.geoservices@ign.fr
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AddressProperties } from './address-properties';
import { GeometryPoint } from './geometry-point';


export interface Address { 
    type?: string;
    properties?: AddressProperties;
    geometry?: GeometryPoint;
}